backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 # Or your specific database driver
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
requests>=2.30.0 # For making requests in next-auth callback, actually needed frontend
PyMySQL>=1.1.0
email-validator>=2.1.0 